backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pillow>=10.4.0
numpy>=2.0.0
opencv-python-headless>=4.10.0
pymupdf>=1.24.3
pytesseract>=0.3.10
google-generativeai>=0.5.2
qrcode[pil]>=7.4.2
